Ajustes no cardapio

+++ src/main/kotlin/br/com/unip/cardapio/domain/campos/CampoMonetario.kt
@@ -1,21 +1,20 @@
 package br.com.unip.cardapio.domain.campos
 
 import br.com.unip.cardapio.exception.CampoNumericoException
-import java.math.BigDecimal
 
-class CampoMonetario : ICampo<BigDecimal> {
+class CampoMonetario : ICampo<Double> {
 
     private val REGEX_NUMEROS: Regex = "-?\\d+(\\.\\d+)?".toRegex()
-    private val valor: BigDecimal
+    private val valor: Double
 
     constructor(valor: ICampo<String>) {
         if (!valor.get().matches(REGEX_NUMEROS)) {
             throw CampoNumericoException()
         }
-        this.valor = BigDecimal(valor.get())
+        this.valor = valor.get().toDouble()
     }
 
-    override fun get(): BigDecimal {
+    override fun get(): Double {
         return valor
     }
 }
